cleanup: Move utils/distribution.py to utils/numeric/distribution (#544)
Setup for: https://github.com/kubernetes-sigs/inference-perf/issues/449

## Summary

Relocates the distribution helpers from
`inference_perf/utils/distribution.py` to
`inference_perf/utils/numeric/distribution/utils.py`. The new package's
`__init__.py` re-exports `generate_distribution` and
`sample_from_distribution`, so the import path is now:

```python
from inference_perf.utils.numeric.distribution import generate_distribution, sample_from_distribution
```

All call sites (datagen modules and tests) are updated accordingly. The
module contents are unchanged (100% rename per git).
