+++ inference_perf/datagen/random_datagen.py
@@ -20,7 +20,7 @@ import numpy as np
 from inference_perf.apis import CompletionAPIData, InferenceAPIData, LazyLoadInferenceAPIData
 from inference_perf.config import APIConfig, APIType, DataConfig, TraceFormat
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
-from inference_perf.utils.distribution import generate_distribution
+from inference_perf.utils.numeric.distribution import generate_distribution
 from inference_perf.utils.trace_reader import AzurePublicDatasetReader
 from .base import DataGenerator, LazyLoadDataMixin
 from .datagen_utils import generate_random_exact_length_text, init_vocab_sampling, random_token_ids
